+++ pecha_api/app.py
@@ -20,7 +20,7 @@ from pecha_api.search import search_views
 from pecha_api.plans.auth import plan_auth_views
 from pecha_api.plans.cms import views as cms_plans_views
 from pecha_api.plans.public import views as public_plans_views
-from pecha_api.plans.user import views as user_plans_views
+from pecha_api.plans.users import views as user_plans_views
 from pecha_api.plans.image_upload import media_views
 import uvicorn
