+++ pecha_api/plans/plans_repository.py
@@ -5,7 +5,7 @@ from sqlalchemy import func, asc, desc
 from typing import Optional
 from ..plans.plans_models import Plan
 from pecha_api.plans.items.plan_items_models import PlanItem
-from pecha_api.plans.users.user_plan_progress_models import UserPlanProgress
+from ..plans.users.models import UserPlanProgress
 from fastapi import HTTPException
 from starlette import status
 from pecha_api.plans.plans_response_models import PlansRepositoryResponse, PlanWithAggregates
