+++ pecha_api/plans/users/service.py
@@ -7,7 +7,7 @@ from starlette import status
 from pecha_api.error_contants import ErrorConstants
 from pecha_api.plans.plans_response_models import PlansResponse
 from pecha_api.plans.shared.utils import load_plans_from_json, convert_plan_model_to_dto
-from pecha_api.plans.user.models import UserPlanProgress, UserPlanEnrollRequest
+from pecha_api.plans.users.models import UserPlanProgress, UserPlanEnrollRequest
 from pecha_api.users.users_service import validate_and_extract_user_details
