+++ pecha_api/plans/users/user_plan_progress_repository.py
@@ -1,7 +1,7 @@
 from uuid import UUID
 from sqlalchemy.exc import IntegrityError
 from sqlalchemy.orm import Session
-from .user_plan_progress_models import UserPlanProgress
+from .models import UserPlanProgress
 from fastapi import HTTPException
 from starlette import status
 from pecha_api.plans.auth.plan_auth_models import ResponseError
