+++ pecha_api/plans/users/views.py
@@ -6,8 +6,8 @@ from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
 from typing import Annotated
 
 from pecha_api.plans.plans_response_models import PlansResponse
-from pecha_api.plans.user.models import UserPlanEnrollRequest, UserPlanProgress, UserPlanProgressUpdate
-from pecha_api.plans.user.service import (
+from pecha_api.plans.users.models import UserPlanEnrollRequest, UserPlanProgress
+from pecha_api.plans.users.service import (
     get_user_enrolled_plans,
     enroll_user_in_plan,
     get_user_plan_progress
