+++ tests/plans/test_plan_respository.py
@@ -7,7 +7,7 @@ from sqlalchemy.orm import sessionmaker
 from pecha_api.db.database import Base
 from pecha_api.plans.plans_models import Plan
 from pecha_api.plans.items.plan_items_models import PlanItem
-from pecha_api.plans.users.user_plan_progress_models import UserPlanProgress
+from pecha_api.plans.users.models import UserPlanProgress
 from pecha_api.plans.authors.plan_author_model import Author
 from pecha_api.plans.plans_repository import save_plan, get_plans
 from pecha_api.plans.plans_response_models import PlansRepositoryResponse, PlanWithAggregates
