+++ tests/plans/test_plan_service.py
@@ -3,18 +3,17 @@ import pytest
 from unittest.mock import patch, MagicMock, ANY
 from fastapi import HTTPException
 
-import pecha_api.plans.plans_service as plans_service
+import pecha_api.plans.cms.service as plans_service
 from pecha_api.plans.plans_enums import DifficultyLevel, PlanStatus
 from pecha_api.plans.plans_models import Plan
 from pecha_api.plans.plans_response_models import (
     CreatePlanRequest, UpdatePlanRequest, PlanStatusUpdate,
-    PlanDTO, TaskDTO, PlanDayDTO, PlanWithDays,
-    PlanWithAggregates, PlansRepositoryResponse
+    PlanDTO,PlanWithAggregates, PlansRepositoryResponse
 )
-from pecha_api.plans.plans_service import (
+from pecha_api.plans.cms.service import (
     create_new_plan, get_filtered_plans, get_details_plan,
     update_plan_details, update_selected_plan_status, delete_selected_plan,
-    DUMMY_PLANS, DUMMY_TASKS, DUMMY_DAYS
+    DUMMY_PLANS, DUMMY_DAYS
 )
 
 
@@ -43,11 +42,11 @@ def test_create_new_plan_success():
     saved_plan.image_url = request.image_url
     saved_plan.status = PlanStatus.DRAFT
 
-    with patch("pecha_api.plans.plans_service.SessionLocal") as mock_session_local, \
-        patch("pecha_api.plans.plans_service.save_plan") as mock_save_plan, \
-        patch("pecha_api.plans.plans_service.save_plan_items") as mock_save_plan_items, \
-        patch("pecha_api.plans.plans_service.get_plan_progress") as mock_get_plan_progress, \
-        patch("pecha_api.plans.plans_service.validate_and_extract_author_details") as mock_validate_author:
+    with patch("pecha_api.plans.cms.service.SessionLocal") as mock_session_local, \
+        patch("pecha_api.plans.cms.service.save_plan") as mock_save_plan, \
+        patch("pecha_api.plans.cms.service.save_plan_items") as mock_save_plan_items, \
+        patch("pecha_api.plans.cms.service.get_plan_progress") as mock_get_plan_progress, \
+        patch("pecha_api.plans.cms.service.validate_and_extract_author_details") as mock_validate_author:
         db_session = _mock_session_local(mock_session_local)
         mock_save_plan.return_value = saved_plan
         # save_plan_items returns the list of saved items; return a list sized to total_days
@@ -129,9 +128,9 @@ async def test_get_filtered_plans_success():
         total=2,
     )
 
-    with patch("pecha_api.plans.plans_service.SessionLocal") as mock_session_local, \
-        patch("pecha_api.plans.plans_service.get_plans") as mock_get_plans, \
-        patch("pecha_api.plans.plans_service.validate_and_extract_author_details") as mock_validate_author:
+    with patch("pecha_api.plans.cms.service.SessionLocal") as mock_session_local, \
+        patch("pecha_api.plans.cms.service.get_plans") as mock_get_plans, \
+        patch("pecha_api.plans.cms.service.validate_and_extract_author_details") as mock_validate_author:
         db_session = _mock_session_local(mock_session_local)
         mock_get_plans.return_value = repo_response
         mock_validate_author.return_value = MagicMock()
@@ -185,7 +184,7 @@ async def test_get_details_plan_success():
     # Create a test plan ID that exists in DUMMY_PLANS
     test_plan = DUMMY_PLANS[0]
     
-    with patch("pecha_api.plans.plans_service.validate_and_extract_author_details") as mock_validate_author:
+    with patch("pecha_api.plans.cms.service.validate_and_extract_author_details") as mock_validate_author:
         mock_validate_author.return_value = MagicMock()
         
         response = await get_details_plan(token="dummy-token", plan_id=test_plan.id)
@@ -202,7 +201,7 @@ async def test_get_details_plan_success():
 async def test_get_details_plan_not_found():
     non_existent_id = uuid.uuid4()
     
-    with patch("pecha_api.plans.plans_service.validate_and_extract_author_details") as mock_validate_author:
+    with patch("pecha_api.plans.cms.service.validate_and_extract_author_details") as mock_validate_author:
         mock_validate_author.return_value = MagicMock()
         
         with pytest.raises(HTTPException) as exc_info:
@@ -224,7 +223,7 @@ async def test_update_plan_details_success():
         image_url="https://example.com/updated.jpg"
     )
     
-    with patch("pecha_api.plans.plans_service.validate_and_extract_author_details") as mock_validate_author:
+    with patch("pecha_api.plans.cms.service.validate_and_extract_author_details") as mock_validate_author:
         mock_validate_author.return_value = MagicMock()
         
         response = await update_plan_details(
@@ -247,7 +246,7 @@ async def test_update_plan_details_not_found():
     non_existent_id = uuid.uuid4()
     update_request = UpdatePlanRequest(title="Updated Title")
     
-    with patch("pecha_api.plans.plans_service.validate_and_extract_author_details") as mock_validate_author:
+    with patch("pecha_api.plans.cms.service.validate_and_extract_author_details") as mock_validate_author:
         mock_validate_author.return_value = MagicMock()
         
         with pytest.raises(HTTPException) as exc_info:
@@ -268,7 +267,7 @@ async def test_update_selected_plan_status_success():
     
     status_update = PlanStatusUpdate(status=PlanStatus.PUBLISHED)
     
-    with patch("pecha_api.plans.plans_service.validate_and_extract_author_details") as mock_validate_author:
+    with patch("pecha_api.plans.cms.service.validate_and_extract_author_details") as mock_validate_author:
         mock_validate_author.return_value = MagicMock()
         
         response = await update_selected_plan_status(
@@ -302,7 +301,7 @@ async def test_update_selected_plan_status_invalid_transition():
 
     status_update = PlanStatusUpdate(status=PlanStatus.PUBLISHED)
     
-    with patch("pecha_api.plans.plans_service.validate_and_extract_author_details") as mock_validate_author:
+    with patch("pecha_api.plans.cms.service.validate_and_extract_author_details") as mock_validate_author:
         mock_validate_author.return_value = MagicMock()
         
         with pytest.raises(HTTPException) as exc_info:
@@ -322,7 +321,7 @@ async def test_delete_selected_plan_success():
     test_plan = DUMMY_PLANS[0]
     initial_plan_count = len(DUMMY_PLANS)
     
-    with patch("pecha_api.plans.plans_service.validate_and_extract_author_details") as mock_validate_author:
+    with patch("pecha_api.plans.cms.service.validate_and_extract_author_details") as mock_validate_author:
         mock_validate_author.return_value = MagicMock()
         
         await delete_selected_plan(token="dummy-token", plan_id=test_plan.id)
